+++ dessia_common/workflow.py
@@ -25,8 +25,6 @@ from dessia_common.vectored_objects import from_csv
 from dessia_common.typings import JsonSerializable, Subclass, MethodType
 import warnings
 
-# import plot_data
-
 # Type Aliases
 VariableTypes = Union['Variable', 'TypedVariable',
                       'VariableWithDefaultValue',
@@ -780,17 +778,17 @@ class MultiPlot(Display):
         values2d = [{key: val[key]} for key in first_vars for val in
                     values]
 
-        tooltip = plot_data.Tooltip(name='Tooltip',
-                                    to_disp_attribute_names=self.attributes)
+        tooltip = plot_data.Tooltip(name='Tooltip', attributes=self.attributes)
 
         scatterplot = plot_data.Scatter(tooltip=tooltip,
-                                        to_disp_attribute_names=first_vars,
+                                        x_variable=first_vars[0],
+                                        y_variable=first_vars[1],
                                         elements=values2d,
                                         name='Scatter Plot')
 
         rgbs = [[192, 11, 11], [14, 192, 11], [11, 11, 192]]
         parallelplot = plot_data.ParallelPlot(
-            disposition='horizontal', to_disp_attribute_names=self.attributes,
+            disposition='horizontal', axes=self.attributes,
             rgbs=rgbs, elements=values
         )
         objects = [scatterplot, parallelplot]
@@ -1233,6 +1231,7 @@ class Workflow(Block):
         }
         jsonschemas['run']['required'] = required_inputs
         jsonschemas['run']['method'] = True
+        jsonschemas['run']['python_typing'] = serialize_typing(MethodType)
         return jsonschemas
 
     def to_dict(self):
@@ -1664,9 +1663,12 @@ class Workflow(Block):
         return data
 
     def plot(self):
-        self.plot_jointjs()
+        self.plot_jointjs(warn=False)
 
-    def plot_jointjs(self):
+    def plot_jointjs(self, warn: bool = True):
+        if warn:
+            warnings.warn("Directly calling plot_jointjs is deprecated.\n"
+                          "Please use plot instead.")
         data = json.dumps(self.jointjs_data())
         rendered_template = workflow_template.substitute(workflow_data=data)