+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unrealeased [=> 0.4.7]
+## Added
+- File typings
+- Support for Iterator typing
+- Object's default_dicts now implement objct_class
+
 ## 0.4.6
 
 ### Added